+++ ProcessOptimizer/tests/test_plot.py
@@ -9,9 +9,13 @@ import json
 import ProcessOptimizer as Optimizer
 from ProcessOptimizer.learning import cook_estimator
 from ProcessOptimizer.learning.gaussian_process.kernels import Matern
-from ProcessOptimizer.plots import plot_Pareto_bokeh
-from ProcessOptimizer.plots import plot_objective
-from ProcessOptimizer.plots import plot_objective_1d
+from ProcessOptimizer.plots import (
+    plot_Pareto_bokeh,
+    plot_objective,
+    plot_objective_1d,
+    plot_brownie_bee_frontend,
+)
+from ProcessOptimizer.space import Categorical
 
 
 def build_Pareto_opt():
@@ -210,6 +214,82 @@ def test_plot_objective_1d():
     fig = axes[0][0].figure
     assert axes.shape == (2, 3)
     assert isinstance(fig, mpl.figure.Figure)
+
+
+def test_plot_brownie_bee():
+    # Integration test of the plot function for Brownie Bee
+    
+    max_q = 10
+    # Helper function to convery y-values to stars
+    def star_score(Y, max_stars, scale_min, scale_max):
+        # Get the scale of the system
+        y_scale = scale_max - scale_min
+        # Map Y into the scale 0 to max_stars
+        score = round((scale_max - Y) / y_scale * max_stars)
+        # If Y is outside the system min/max values (can happen due to noise)
+        # we clamp to the scale.
+        score = min(max(score, 0), max_stars)        
+        return score
+    
+    space = [
+        (15.0, 25.0),
+        (70.0, 90.0),
+        (0.03, 0.15),
+        (70.0, 167.0),
+        ["A", "B"],
+    ]
+    opt = Optimizer.Optimizer(
+        space,
+        acq_func="EI",
+        n_initial_points=5,
+    )
+    x = [
+        [18, 84, 0.042, 102.8, "A"],
+        [24, 80, 0.09, 102.8, "B"],
+        [20, 76, 0.066, 102.8, "A"],
+        [16, 72, 0.111, 102.8, "A"],
+        [22, 88, 0.138, 102.8, "B"],
+        [24.5, 70, 0.145, 102.8, "A"],
+        [19, 79, 0.124, 102.8, "B"],
+        [15.5, 80, 0.100, 106, "B"],
+        [24, 86, 0.111, 106, "B"],
+        [18, 79, 0.124, 109, "A"],
+        [16, 75, 0.145, 109, "B"],
+        [17.9, 74, 0.150, 122.9, "A"],
+        [22.6, 89, 0.075, 122.9, "B"],
+        [23, 70, 0.145, 118.6, "A"],
+        [23, 84, 0.042, 118.6, "B"],
+    ]
+    y = [54, 53, 38, 30, 65, 42, 60, 77, 66, 75, 61, 57, 87, 47, 67]
+    y_star = [-star_score(yy, max_q, 0, 100) for yy in y]
+    res = opt.tell(x, y_star)
+    fig_list = plot_brownie_bee_frontend(res, max_quality=max_q)
+    
+    # Check we built a list of figures
+    for fig in fig_list:
+        assert isinstance(fig, mpl.figure.Figure)
+    # Check that we return the right number of figures in total
+    assert len(fig_list) == len(space)+1
+   
+    # Check that the y-axis uses the correct limits
+    limits = fig_list[0].gca().get_ylim()
+    assert (limits[0] == 0) & (limits[1] == max_q*1.02)
+    
+    # Check that each x-axis of the factor plots uses the correct limits
+    for i, limits in enumerate(space):
+        xlim = fig_list[i].gca().get_xlim()
+        # Categorical factors are special in their encoding
+        if isinstance(opt.space.dimensions[i], Categorical):
+            assert xlim[0] == -0.2
+            assert xlim[1] == len(opt.space.dimensions[i].categories)-0.8
+        else:
+            # Numeric factors should simply map to the space bounds
+            assert (xlim[0] == space[i][0]) & (xlim[1] == space[i][1])
+    
+    # Check that the histogram of expected outputs uses a sensible x-axis
+    xlim = fig_list[-1].gca().get_xlim()
+    assert (xlim[0] >= 0) & (xlim[1] <= max_q*1.02)
+    
 
 
 def test_plot_Pareto_bokeh_return_data():